+++ Air_quality_query/myEmailKey.txt
@@ -1,0 +1,2 @@
+zhu.tim99@gmail.com
+copperswift75
