--- Guide_to_get_APIkey.md
@@ -1,10 +1,0 @@
-
-#### For Air Quality Data (AQS) ~~
-
-Check this Website for details
-`https://aqs.epa.gov/aqsweb/documents/data_api.html`
-
-Generally, the return you get from the API would in the form of **json** ~
-
-To sign up for the API key for the AQS
-`https://aqs.epa.gov/data/api/signup?email={Your email here}`
