--- myEmailKey.txt
@@ -1,2 +1,0 @@
-zhu.tim99@gmail.com
-copperswift75
